+++ src/tests/createStore.test.js
@@ -1,6 +1,6 @@
 import { getStore } from "../index";
 
-it("injectReducer key must be unique", () => {
-  getStore().injectReducer("myKey", () => 1);
-  getStore().injectReducer("myKey", () => 2); // second call OK
+it("registerReducers with the same name overrides - for Hot-reloading support", () => {
+  getStore().registerReducers({"my-action-type": () => 1});
+  getStore().registerReducers({"my-action-type": () => 2});
 });