+++ src/tests/storeRegistry.test.js
@@ -2,6 +2,6 @@ import { getStore, setStore, createStore } from "../index";
 
 it("works", () => {
   const store1 = getStore()
-  const store2 = setStore(createStore({}))
+  const store2 = setStore(createStore())
   expect(store1 === store2).toEqual(false)
 });